Check if file exists before linking

+++ scripts/create_links_local_dotfiles.sh
@@ -2,14 +2,13 @@
 
 # NOTE: Must be executed from local dotfiles directory
 
-ln --symbolic --interactive --verbose configs/profile.local "$HOME"/.profile.local
+[[ -f configs/profile.local ]] && ln --symbolic --interactive --verbose configs/profile.local "$HOME"/.profile.local
 
-ln --symbolic --interactive --verbose configs/bash_profile.local "$HOME"/.bash_profile.local
-ln --symbolic --interactive --verbose configs/bashrc.local "$HOME"/.bashrc.local
-ln --symbolic --interactive --verbose configs/bash_aliases.local "$HOME"/.bash_aliases.local
+[[ -f configs/bashrc.local ]] && ln --symbolic --interactive --verbose configs/bashrc.local "$HOME"/.bashrc.local
+[[ -f configs/bash_aliases.local ]] && ln --symbolic --interactive --verbose configs/bash_aliases.local "$HOME"/.bash_aliases.local
 
-ln --symbolic --interactive --verbose configs/gitconfig.local "$HOME"/.gitconfig.local
+[[ -f configs/gitconfig.local ]] && ln --symbolic --interactive --verbose configs/gitconfig.local "$HOME"/.gitconfig.local
 
-ln --symbolic --interactive --verbose configs/tmux.conf.local "$HOME"/.tmux.conf.local
+[[ -f configs/tmux.conf.local ]] && ln --symbolic --interactive --verbose configs/tmux.conf.local "$HOME"/.tmux.conf.local
 
-ln --symbolic --interactive --verbose configs/vimrc.local "$HOME"/.vimrc.local
+[[ -f configs/vimrc.local ]] && ln --symbolic --interactive --verbose configs/vimrc.local "$HOME"/.vimrc.local
